+++ app/hospital/[hos_id]/icu/[target_date]/search/page.tsx
@@ -1,9 +1,9 @@
-import IcuSearchChart from '@/components/hospital/icu/main/search/icu-search-chart'
+import IcuSearchChartEntry from '@/components/hospital/icu/main/search/icu-search-chart-entry'
 
 export default function SearchPage() {
   return (
     <div className="mt-12 flex h-mobile flex-col border-t p-2 2xl:mt-0 2xl:h-desktop 2xl:border-0">
-      <IcuSearchChart />
+      <IcuSearchChartEntry />
     </div>
   )
 }
