+++ app/hospital/[hos_id]/icu/[target_date]/template/page.tsx
@@ -1,4 +1,4 @@
-import Template from '@/components/hospital/icu/main/template/template'
+import IcuTemplateEntry from '@/components/hospital/icu/main/template/icu-template-entry'
 import { getTemplateCharts } from '@/lib/services/icu/template/template'
 
 type TemplatePageProps = {
@@ -13,7 +13,7 @@ export default async function TemplatePage(props: TemplatePageProps) {
 
   return (
     <div className="mt-12 flex h-mobile flex-col border-t p-2 2xl:mt-0 2xl:h-desktop 2xl:border-0">
-      <Template templateCharts={templateCharts} />
+      <IcuTemplateEntry templateCharts={templateCharts} />
     </div>
   )
 }
