+++ components/hospital/icu/main/search/sheet/search-type-radio.tsx
@@ -1,4 +1,4 @@
-import type { SearchOptions } from '@/components/hospital/icu/main/search/icu-search-chart'
+import type { SearchOptions } from '@/components/hospital/icu/main/search/icu-search-chart-entry'
 import { Label } from '@/components/ui/label'
 import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group'
 import type { Dispatch, SetStateAction } from 'react'
